+++ src/main/java/com/lhy/wcwt/client/ModClientSetup.java
@@ -62,7 +62,6 @@ public class ModClientSetup {
             event.setCanceled(true);
             return;
         }
-
         if (screen == null) {
             return;
         }